# src/app/lib/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Marker Web API",
    "version": "0.4.0",
    "description": "OpenAPI documentation for the Marker Web Next.js route handlers. Keep this file synchronized with src/app/api whenever endpoints change."
  },  "servers": [
    {
      "url": "/",
      "description": "Current Next.js server"
    }
  ],
...
  ],  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "marker_session",
        "description": "Opaque HttpOnly, Secure, SameSite=Strict session cookie. Set by login and never exposed to JavaScript."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "ObjectId": {
        "type": "string",
        "example": "6652f4b88f0f3c7101111111"
      },
      "AuthLoginRequest": {
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
...
      "AdminLoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "age": {
            "oneOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ]
          },
          "sex": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string"
          },
          "connections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          }
        }
      },
...
      "EventInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPrivate": {
            "type": "boolean"
          },
          "media": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded. Retry after the number of seconds in Retry-After.",
        "headers": {
          "Retry-After": {
            "description": "Seconds until the current rate-limit window resets.",
            "schema": {
              "type": "integer"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "Request body exceeds the endpoint size limit."
      }
    }
  },
  "paths": {
...
    "/api/admin/auth/login": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Login as an administrator",
        "description": "Authenticates a database user whose role is admin and sets the same opaque session cookie.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
...
    "/api/user/get-user-collection": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Get users by id collection",
        "description": "Fetches multiple users by ids. This route currently does not enforce auth.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ObjectId"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Users loaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/user/search-users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Search users",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Users loaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
...
    "/api/user/unfollow-user": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Unfollow a user",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "$ref": "#/components/schemas/ObjectId"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unfollowed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
...
    "/api/event/get-user-events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get legacy events",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "ObjectId cursor returned as nextCursor by the previous page.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events loaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/EventInput"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }}